\name{lproj-package}
\alias{lproj-package}
\alias{lproj}
\docType{package}
\title{
\packageTitle{lproj}
}
\description{
\packageDescription{lproj}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lproj}
\packageIndices{lproj}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{lproj}

Maintainer: \packageMaintainer{lproj}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
